+++ deploy/web.sh
@@ -1,0 +1,2 @@
+flutter build web --dart-define-from-file=.secrets.json.firebase --web-renderer=html
+firebase deploy