+++ lib/providers/state.dart
@@ -1,5 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 import 'package:poll_and_play/config.dart';
 import 'package:poll_and_play/grpc/user.dart';
 import 'package:poll_play_proto_gen/public.dart' as proto;
@@ -9,11 +10,13 @@ import 'package:poll_play_proto_gen/public.dart' as proto;
 class StateProvider extends ChangeNotifier {
   final UserClient _client = UserClient(GlobalConfig().apiAddress.split(':'));
   proto.User? _user;
+  late PackageInfo packageInfo;
 
   StateProvider();
 
   Future<void> init() async {
     await initUser();
+    packageInfo = await PackageInfo.fromPlatform();
   }
 
   Future<void> initUser() async {